removed appsettings and added user secrets; css not rendering on page

+++ AsyncInn/AsyncInn/Models/HotelRoom.cs
@@ -11,9 +11,9 @@ namespace AsyncInn.Models
 
         public int RoomNumber { get; set; }
 
-        public decimal RoomID { get; set; }
+        public int RoomID { get; set; }
 
-        public decimal Rate { get; set; }
+        public int Rate { get; set; }
 
         public bool PetFriendly { get; set; }
